requirements.txt
vosk==0.3.45
sounddevice==0.4.5
numpy>=1.19.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pyttsx3>=2.90
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
